Optimizing code and organizing code
