+++ src/components/Scroll.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Scroll = (props) =>  {
+	return (
+		<div style = {{ overflowY: 'scroll', borderTop: '2px solid black', height: '500px' }}>  
+			{props.children}
+		</div>
+	);
+};
+
+export default Scroll;
